Added wall sound collider to car pawn, change FOV behavior
A wall sound collider has been placed on the car pawn to allow easier implementation of wall hit sounds. Both cameras have also had an FOV change to 100 degrees and there is now a functionality in place that zooms FOV out up to 110 degrees while boosting to help display a sense of speed.